// src/app/modules/citas-medicas/models/hora.ts
export interface Hora {
  consultorioHorarioId: number
  descripcion: string
  descripcionInhabilitacion: string
  estado: boolean
  fin: string
  finDetalle: string
  horarioId: number
  horarioValido: boolean
  id: number
  inhabilitado: boolean
  inicio: string
  inicioDetalle: string
  isAdicional: boolean
  numero: number
}